# Add --enable-trace-log install option