--- a/ext/oj/extconf.rb
+++ b/ext/oj/extconf.rb
@@ -39,6 +39,10 @@ if with_config('--with-sse42')
   dflags['OJ_USE_SSE4_2'] = 1
 end
 
+if enable_config('trace-log', false)
+  dflags['OJ_ENABLE_TRACE_LOG'] = 1
+end
+
 dflags.each do |k,v|
   if v.nil?
     $CPPFLAGS += " -D#{k}"