I think the trace log feature is for debugging and most people don't use it. Therefore, I usually want to improve performance by removing unnecessary instructions.

−       | before | after  | result
--       | --     | --     | --
Oj.dump  | 1.440M | 1.697M | 1.178x

### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 6.1.1-1-MANJARO
  - AMD Ryzen 7 5800H
  - gcc version 12.2.0
  - Ruby 3.2.0

### Before
```
Warming up --------------------------------------
             Oj.dump   142.510k i/100ms
Calculating -------------------------------------
             Oj.dump      1.440M (± 0.4%) i/s -     14.536M in  10.091795s
```

### After
```
Warming up --------------------------------------
             Oj.dump   165.140k i/100ms
Calculating -------------------------------------
             Oj.dump      1.697M (± 0.7%) i/s -     17.009M in  10.025122s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

data = [[1,2,3,4,5]]*10

Benchmark.ips do |x|
  x.time = 10

  x.report('Oj.dump') do
    Oj.dump(data)
  end
end
```