@ohler55 2023-01-08
I agree, it is rarely used to having a compile options makes sense.